# Cleanup and performance improvements