--- a/Fluid.MvcViewEngine/FluidMvcViewOptionsSetup.cs
+++ b/Fluid.MvcViewEngine/FluidMvcViewOptionsSetup.cs
@@ -13,8 +13,11 @@ namespace Fluid.MvcViewEngine
         /// <param name="fluidViewEngine">The <see cref="IFluidViewEngine"/>.</param>
         public FluidMvcViewOptionsSetup(IFluidViewEngine fluidViewEngine)
         {
-            _fluidViewEngine = fluidViewEngine
-                               ?? ExceptionHelper.ThrowArgumentNullException<IFluidViewEngine>(nameof(fluidViewEngine));
+            if (fluidViewEngine is null)
+            {
+                ExceptionHelper.ThrowArgumentNullException(nameof(fluidViewEngine));
+            }
+            _fluidViewEngine = fluidViewEngine;
         }
 
         /// <summary>
@@ -23,10 +26,9 @@ namespace Fluid.MvcViewEngine
         /// <param name="options">The <see cref="MvcViewOptions"/> to configure.</param>
         public void Configure(MvcViewOptions options)
         {
-            if (options == null)
+            if (options is null)
             {
                 ExceptionHelper.ThrowArgumentNullException(nameof(options));
-                return;
             }
 
             options.ViewEngines.Add(_fluidViewEngine);