--- a/Fluid/Ast/IfStatement.cs
+++ b/Fluid/Ast/IfStatement.cs
@@ -15,7 +15,7 @@ namespace Fluid.Ast
             List<Statement> statements,
             ElseStatement elseStatement = null,
             List<ElseIfStatement> elseIfStatements = null
-            ) :base (statements)
+        ) : base(statements)
         {
             Condition = condition;
             Else = elseStatement;