--- a/Fluid/Ast/TagStatement.cs
+++ b/Fluid/Ast/TagStatement.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace Fluid.Ast
 {