--- a/Fluid/ExceptionHelper.cs
+++ b/Fluid/ExceptionHelper.cs
@@ -1,26 +1,37 @@
+#nullable enable
+
 using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 
 namespace Fluid
 {
     internal static class ExceptionHelper
     {
+        [DoesNotReturn]
         [MethodImpl(MethodImplOptions.NoInlining)]
-        public static void ThrowArgumentNullException(string paramName)
+        public static void ThrowArgumentNullException(string paramName, string? message = null)
         {
-            ThrowArgumentNullException<object>(paramName);
+            throw new ArgumentNullException(paramName, message);
         }
 
+        [DoesNotReturn]
         [MethodImpl(MethodImplOptions.NoInlining)]
-        public static T ThrowArgumentNullException<T>(string paramName)
+        public static void ThrowArgumentOutOfRangeException(string paramName, string message)
         {
-            throw new ArgumentNullException(paramName);
+            throw new ArgumentOutOfRangeException(paramName, message);
         }
 
+        [DoesNotReturn]
         [MethodImpl(MethodImplOptions.NoInlining)]
-        public static T ThrowParseException<T>(string message)
+        public static void ThrowParseException<T>(string message)
         {
             throw new ParseException(message);
         }
+
+#if NETSTANDARD2_0
+        private class DoesNotReturnAttribute : Attribute {}
+#endif
+
     }
 }