I was a bit annoyed by the perceived render performance shown in benchmarks, so I went hunting and fixed things along the way. The fixes should make pretty much sense in review-sense, many of them are not touched by the current benchmarks though.

* reduce LINQ usage
* use pre-allocated objects for integers and chars
* remove some interface dispatch
* make member access strategy abstract class instead of interface
* check some common paths that can be optimized
* add some internal FluidValue helpers for scenarios like ToList, ToArray, FirstOrDefault which are optimized per underlying type

## Fluid.Benchmarks.FluidBenchmarks

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Parse|8.529 μs|0.1550 μs|0.6714|0.0000|0.0000|2.77 KB|
| **New** |	| **8.166 μs (-4%)** | **0.1522 μs** | **0.6256 (-7%)** | **0.0000** | **0.0000** | **2.57 KB (-7%)** |
| Old |ParseBig|48.476 μs|0.6396 μs|2.8076|0.0000|0.0000|11.63 KB|
| **New** |	| **48.516 μs (0%)** | **0.7750 μs** | **2.8076 (0%)** | **0.1831** | **0.0000** | **11.5 KB (-1%)** |
| Old |Render|787.805 μs|7.7146 μs|90.8203|58.5938|30.2734|390.76 KB|
| **New** |	| **704.018 μs (-11%)** | **2.7758 μs** | **90.8203 (0%)** | **58.5938 (0%)** | **30.2734 (0%)** | **370.91 KB (-5%)** |
| Old |ParseAndRender|801.069 μs|11.9312 μs|90.8203|58.5938|30.2734|393.66 KB|
| **New** |	| **717.019 μs (-10%)** | **2.4128 μs** | **90.8203 (0%)** | **58.5938 (0%)** | **30.2734 (0%)** | **373.6 KB (-5%)** |