@sebastienros 2021-01-23
That's great, thanks.

Next step, a Compile() method that will generate a compiled renderer, and ultimately compile automatically a template in the background that is called many times.

Will probably do it first in Parlot (explicitly called) based on the scanner api. Like regexes can do.